requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
colorama==0.3.6           # via pytest
py==1.4.31                # via pytest
pytest==2.8.7
python-dateutil==2.5.0
requests==2.9.1
six==1.10.0               # via python-dateutil
